+++ backend/systembridgebackend/modules/battery/update.py
@@ -1,8 +1,8 @@
 """System Bridge: Update Battery"""
-from systembridgebackend.common import camel_to_snake
-from systembridgebackend.database import Database
 from systembridgebackend.modules.base import ModuleUpdateBase
 from systembridgebackend.modules.battery import Battery
+from systembridgeshared.common import camel_to_snake
+from systembridgeshared.database import Database
 
 
 class BatteryUpdate(ModuleUpdateBase):
