+++ backend/systembridgebackend/modules/cpu/update.py
@@ -1,5 +1,5 @@
 """System Bridge: Update CPU"""
-from systembridgebackend.database import Database
+from systembridgeshared.database import Database
 from systembridgebackend.modules.base import ModuleUpdateBase
 from systembridgebackend.modules.cpu import CPU
