+++ backend/systembridgebackend/modules/disk/update.py
@@ -1,5 +1,5 @@
 """System Bridge: Update Disk"""
-from systembridgebackend.database import Database
+from systembridgeshared.database import Database
 from systembridgebackend.modules.base import ModuleUpdateBase
 from systembridgebackend.modules.disk import Disk
