+++ backend/systembridgebackend/modules/memory/__init__.py
@@ -3,7 +3,7 @@ from collections import namedtuple
 
 from psutil import swap_memory, virtual_memory
 from psutil._common import sswap
-from systembridgebackend import Base
+from systembridgeshared.base import Base
 
 
 class Memory(Base):
