+++ backend/systembridgebackend/modules/network/__init__.py
@@ -1,7 +1,7 @@
 """System Bridge: Network"""
 from psutil import net_connections, net_if_addrs, net_if_stats, net_io_counters
 from psutil._common import sconn, snetio, snicaddr, snicstats
-from systembridgebackend import Base
+from systembridgeshared.base import Base
 
 
 class Network(Base):
