+++ backend/systembridgebackend/modules/network/update.py
@@ -1,5 +1,5 @@
 """System Bridge: Update Network"""
-from systembridgebackend.database import Database
+from systembridgeshared.database import Database
 from systembridgebackend.modules.base import ModuleUpdateBase
 from systembridgebackend.modules.network import Network
