+++ backend/systembridgebackend/modules/sensors/update.py
@@ -1,5 +1,5 @@
 """System Bridge: Update Sensors"""
-from systembridgebackend.database import Database
+from systembridgeshared.database import Database
 from systembridgebackend.modules.base import ModuleUpdateBase
 from systembridgebackend.modules.sensors import Sensors
