+++ backend/systembridgebackend/modules/update.py
@@ -1,8 +1,8 @@
 """System Bridge: Modules Update"""
 import asyncio
 
-from systembridgebackend import Base
-from systembridgebackend.database import Database
+from systembridgeshared.base import Base
+from systembridgeshared.database import Database
 from systembridgebackend.modules.battery.update import BatteryUpdate
 from systembridgebackend.modules.bridge.update import BridgeUpdate
 from systembridgebackend.modules.cpu.update import CPUUpdate
