+++ backend/systembridgebackend/server/auth.py
@@ -2,7 +2,7 @@
 from functools import wraps
 
 from sanic import Sanic, exceptions
-from systembridgebackend import Base
+from systembridgeshared.base import Base
 
 
 class ApiKeyAuthentication(Base):
