+++ cli/requirements.txt
@@ -1,7 +1,3 @@
-appdirs==1.4.4
-colorama==0.4.4
-cryptography==36.0.2
-pandas==1.4.2
 pyinstaller==4.10
 shellingham==1.3.1
 tabulate==0.8.9
