+++ cli/systembridgecli/__main__.py
@@ -2,13 +2,10 @@
 import typer
 from tabulate import tabulate
 
-from systembridgecli.database import Database
-from systembridgecli.settings import (
-    SECRET_API_KEY,
-    SETTING_PORT_API,
-    TABLE_SETTINGS,
-    Settings,
-)
+from systembridgeshared.const import SECRET_API_KEY, SETTING_PORT_API, TABLE_SETTINGS
+from systembridgeshared.database import Database
+from systembridgeshared.settings import Settings
+
 
 app = typer.Typer()
 database = Database()
