--- cli/systembridgecli/common.py
@@ -1,5 +1,0 @@
-"""System Bridge CLI: Common"""
-
-COLUMN_KEY = "key"
-COLUMN_VALUE = "value"
-COLUMN_TIMESTAMP = "timestamp"
