+++ gui/requirements.txt
@@ -1,4 +1,4 @@
 plyer==2.0.0
 pyinstaller==4.10
 PySide6==6.2.4
-systembridge==2.3.1
+systembridgeshared==3.0.0
