+++ shared/.vscode/launch.json
@@ -1,0 +1,13 @@
+{
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "name": "Python: Module",
+      "type": "python",
+      "request": "launch",
+      "module": "systembridgecli",
+      "preLaunchTask": "Python: pip install",
+      "args": ["--help"]
+    }
+  ]
+}
