+++ shared/.vscode/tasks.json
@@ -1,0 +1,10 @@
+{
+  "version": "2.0.0",
+  "tasks": [
+    {
+      "type": "shell",
+      "label": "Python: pip install",
+      "command": "pip install ."
+    }
+  ]
+}
