+++ shared/requirements.txt
@@ -1,0 +1,4 @@
+appdirs==1.4.4
+cryptography==36.0.2
+pandas==1.4.2
+pyinstaller==4.10
