+++ shared/systembridgeshared/common.py
@@ -1,10 +1,17 @@
-"""System Bridge: Common"""
+"""System Bridge Shared: Common"""
 from __future__ import annotations
+import os
 import re
+from appdirs import AppDirs
 
-COLUMN_KEY = "key"
-COLUMN_VALUE = "value"
-COLUMN_TIMESTAMP = "timestamp"
+
+def get_user_data_directory() -> str:
+    """Get the user data directory"""
+    user_data_dir = AppDirs("systembridge", "timmo001").user_data_dir
+    # Create User Data Directories
+    if not os.path.exists(user_data_dir):
+        os.makedirs(user_data_dir, exist_ok=True)
+    return user_data_dir
 
 
 def camel_to_snake(name):
