+++ shared/systembridgeshared/const.py
@@ -1,0 +1,20 @@
+"""System Bridge: Constants"""
+
+# Logging
+DATE_FORMAT = "%Y-%m-%d %H:%M:%S"
+FORMAT = "%(asctime)s %(levelname)s (%(threadName)s) [%(name)s] %(message)s"
+
+# Database
+TABLE_SECRETS = "secrets"
+TABLE_SETTINGS = "settings"
+
+COLUMN_KEY = "key"
+COLUMN_VALUE = "value"
+COLUMN_TIMESTAMP = "timestamp"
+
+# Settings
+SETTING_LOG_LEVEL = "log_level"
+SETTING_PORT_API = "port_api"
+
+# Secrets
+SECRET_API_KEY = "api_key"
