+++ shared/version.txt
@@ -1,0 +1,1 @@
+3.0.0
